housework-api/pyproject.toml
[tool.poetry]
name = "housework-api"
version = "0.1.0"
description = ""
authors = ["Tom Newport <tom.newport@gmail.com>"]
readme = "README.md"
packages = [{include = "housework_api"}]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.2.7"
django-ninja = "^0.22.2"
pydantic = {version = "1.10.13", extras = ["email"]}
freezegun = "^1.2.2"
coverage = "^7.3.2"
requests = "^2.31.0"
black = "^23.11.0"
django-cors-headers = "^4.3.0"
pywebpush = "^1.14.0"
celery = "^5.3.5"
redis = "^5.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
